[Bronze V] Title: 이상한 기호, Time: 0 ms, Memory: 1112 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.c
@@ -3,17 +3,17 @@
 #include <stdlib.h>
 #include <string.h>
 
-void C(int A, int B)
+void C(double A, double B)
 {
     
-    printf("%d", (A + B) * (A - B));
+    printf("%.0f", (A + B) * (A - B));
 
 }
 int main() 
 {
-    int A;
-    int B;
-    scanf("%d %d", &A, &B);
+    double A;
+    double B;
+    scanf("%lf %lf", &A, &B);
     C(A,B);
     return 0;
 }
